upstream: [pkg-alt] reorganize the files in the dependency module (#22509)
