+++ external-crates/move/crates/move-package-alt/src/dependency/mod.rs
@@ -3,71 +3,25 @@
 // Modifications Copyright (c) 2025 IOTA Stiftung
 // SPDX-License-Identifier: Apache-2.0
 
+mod combine;
+pub use combine::CombinedDependency;
+
+mod resolve;
+pub use resolve::{ResolvedDependency, ResolverError};
+
+mod pin;
+pub use pin::{PinnedDependencyInfo, pin};
+
+mod fetch;
+pub use fetch::FetchedDependency;
+
 mod dependency_set;
-pub mod external;
-pub mod git;
-pub mod local;
-mod onchain;
-
-use std::{collections::BTreeMap, path::PathBuf};
-
 pub use dependency_set::DependencySet;
-use derive_where::derive_where;
-use futures::future::join_all;
-use git::PinnedGitDependency;
-use serde::{
-    Deserialize, Deserializer, Serialize,
-    de::{self, DeserializeOwned},
-};
-use tokio::join;
-use toml_edit::TomlError;
-use tracing::debug;
 
 use crate::{
-    errors::{FileHandle, PackageResult},
-    flavor::MoveFlavor,
-    git::GitTree,
-    package::{EnvironmentName, manifest::ManifestResult, paths::PackagePath},
-    schema::{
-        Address, DefaultDependency, EnvironmentID, LocalDepInfo, LockfileDependencyInfo,
-        LockfileGitDepInfo, ManifestDependencyInfo, ManifestGitDependency, OnChainDepInfo, Pin,
-        ReplacementDependency, ResolverDependencyInfo,
-    },
+    errors::FileHandle,
+    schema::{Address, EnvironmentName},
 };
-
-#[derive(Debug, Clone)]
-pub struct CombinedDependency(Dependency<Combined>);
-pub struct ResolvedDependency(Dependency<Resolved>);
-pub struct PinnedDependency(Dependency<Pinned>);
-pub struct FetchedDependency(Dependency<Fetched>);
-
-#[derive(Debug, Clone)]
-pub struct PinnedDependencyInfo(Dependency<Pinned>);
-
-/// [Dependency<Combined>]s contain the dependency-type-specific things that
-/// users write in their Move.toml files. They are formed by combining the
-/// entries from the `[dependencies]` and the `[dep-replacements]` section of
-/// the manifest.
-type Combined = ManifestDependencyInfo;
-
-/// A [Dependency<Resolved>] is like a [Dependency<Combined>] except that it no
-/// longer has externally resolved dependencies
-type Resolved = ResolverDependencyInfo;
-
-/// [Dependency<Pinned>]s are guaranteed to always resolve to the same package
-/// source. For example, a git dependendency with a branch or tag revision may
-/// change over time (and is thus not pinned), whereas a git dependency with a
-/// sha revision is always guaranteed to produce the same files.
-#[derive(Clone, Debug)]
-enum Pinned {
-    Local(LocalDepInfo),
-    Git(PinnedGitDependency),
-    OnChain(OnChainDepInfo),
-}
-
-/// Once a dependency has been fetched, it is simply represented by a
-/// [PackagePath]
-type Fetched = PackagePath;
 
 /// [Dependency] wraps information about the location of a dependency (such as
 /// the `git` or `local` fields) with additional metadata about how the
@@ -114,192 +68,4 @@ impl<T> Dependency<T> {
     pub fn use_environment(&self) -> &EnvironmentName {
         &self.use_environment
     }
-
-    pub fn is_override(&self) -> bool {
-        self.is_override
-    }
 }
-impl CombinedDependency {
-    /// Specialize an entry in the `[dependencies]` section, for the environment
-    /// named `source_env_name`
-    pub fn from_default(
-        file: FileHandle,
-        source_env_name: EnvironmentName,
-        default: DefaultDependency,
-    ) -> Self {
-        CombinedDependency(Dependency {
-            dep_info: default.dependency_info,
-            use_environment: source_env_name,
-            is_override: default.is_override,
-            published_at: None,
-            containing_file: file,
-        })
-    }
-
-    /// Load from an entry in the `[dep-replacements]` section that has no
-    /// corresponding entry in the `[dependencies]` section of the manifest.
-    /// `source_env_name` refers to the environment name and ID in the
-    /// original manifest; it is used as the default environment for the
-    /// dependency, but will be overridden if `replacement` specifies
-    /// `use-environment` field.
-    pub fn from_replacement(
-        file: FileHandle,
-        source_env_name: EnvironmentName,
-        replacement: ReplacementDependency,
-    ) -> ManifestResult<Self> {
-        let Some(dep) = replacement.dependency else {
-            return Err(todo!());
-        };
-
-        Ok(CombinedDependency(Dependency {
-            dep_info: dep.dependency_info,
-            use_environment: replacement.use_environment.unwrap_or(source_env_name),
-            is_override: dep.is_override,
-            published_at: replacement.published_at,
-            containing_file: file,
-        }))
-    }
-
-    pub fn from_default_with_replacement(
-        file: FileHandle,
-        source_env_name: EnvironmentName,
-        default: DefaultDependency,
-        replacement: ReplacementDependency,
-    ) -> ManifestResult<Self> {
-        let dep = replacement.dependency.unwrap_or(default);
-
-        // TODO: possibly additional compatibility checks here?
-
-        Ok(CombinedDependency(Dependency {
-            dep_info: dep.dependency_info,
-            use_environment: replacement.use_environment.unwrap_or(source_env_name),
-            is_override: dep.is_override,
-            published_at: replacement.published_at,
-            containing_file: file,
-        }))
-    }
-}
-
-impl PinnedDependencyInfo {
-    /// Return a dependency representing the root package
-    pub fn root_dependency(containing_file: FileHandle, use_environment: EnvironmentName) -> Self {
-        PinnedDependencyInfo(Dependency {
-            dep_info: Pinned::Local(todo!()),
-            use_environment,
-            is_override: true,
-            published_at: None,
-            containing_file,
-        })
-    }
-
-    pub fn from_pin(containing_file: FileHandle, env: &EnvironmentName, pin: &Pin) -> Self {
-        let dep_info = match &pin.source {
-            LockfileDependencyInfo::Local(loc) => Pinned::Local(loc.clone()),
-            LockfileDependencyInfo::OnChain(chain) => Pinned::OnChain(chain.clone()),
-            LockfileDependencyInfo::Git(git) => todo!(),
-        };
-
-        PinnedDependencyInfo(Dependency {
-            dep_info,
-            use_environment: pin.use_environment.clone().unwrap_or(env.clone()),
-            is_override: false, // TODO
-            published_at: None, // TODO
-            containing_file,
-        })
-    }
-
-    pub async fn fetch(&self) -> PackageResult<PathBuf> {
-        // TODO: need to actually fetch local dep
-        match &self.0.dep_info {
-            Pinned::Git(dep) => Ok(dep.fetch().await?),
-            Pinned::Local(dep) => Ok(dep.absolute_path(self.0.containing_file.path())),
-            Pinned::OnChain(dep) => todo!(),
-        }
-    }
-
-    /// Return the absolute path to the directory that this package would be
-    /// fetched into, without actually fetching it
-    /// Return the absolute path to the directory that this package would be
-    /// fetched into, without actually fetching it
-    pub fn unfetched_path(&self) -> PathBuf {
-        match &self.0.dep_info {
-            Pinned::Git(dep) => dep.unfetched_path(),
-            Pinned::Local(dep) => dep.absolute_path(self.0.containing_file.path()),
-            Pinned::OnChain(dep) => todo!(),
-        }
-    }
-
-    pub fn use_environment(&self) -> &EnvironmentName {
-        self.0.use_environment()
-    }
-}
-impl From<PinnedDependencyInfo> for LockfileDependencyInfo {
-    fn from(value: PinnedDependencyInfo) -> Self {
-        match value.0.dep_info {
-            Pinned::Local(loc) => Self::Local(loc),
-            Pinned::Git(git) => Self::Git(LockfileGitDepInfo {
-                repo: git.inner().repo_url().to_string(),
-                rev: git.inner().sha().clone(),
-                path: git.inner().path_in_repo().to_path_buf(),
-            }),
-            Pinned::OnChain(on_chain) => Self::OnChain(on_chain),
-        }
-    }
-}
-
-/// Replace all dependencies with their pinned versions.
-pub async fn pin<F: MoveFlavor>(
-    mut deps: DependencySet<CombinedDependency>,
-    envs: &BTreeMap<EnvironmentName, EnvironmentID>,
-) -> PackageResult<DependencySet<PinnedDependencyInfo>> {
-    use Pinned as P;
-
-    // resolution
-    let deps = ResolvedDependency::resolve(deps, envs).await?;
-    debug!("done resolving");
-
-    // pinning
-    let mut result: DependencySet<PinnedDependencyInfo> = DependencySet::new();
-    for (env, pkg, dep) in deps.into_iter() {
-        let transformed = match &dep.0.dep_info {
-            ResolverDependencyInfo::Local(loc) => P::Local(loc.clone()),
-            ResolverDependencyInfo::Git(git) => P::Git(git.pin().await?),
-            ResolverDependencyInfo::OnChain(chain) => P::OnChain(todo!()),
-        };
-        result.insert(env, pkg, PinnedDependencyInfo(dep.0.map(|_| transformed)));
-    }
-
-    Ok(result)
-}
-
-/// For each environment, if none of the implicit dependencies are present in
-/// [deps] (or the default environment), then they are all added.
-// TODO: what's the notion of identity used here? I think it has to be by name
-fn add_implicit_deps<F: MoveFlavor>(
-    flavor: &F,
-    deps: &mut DependencySet<Dependency<Pinned>>,
-) -> PackageResult<()> {
-    todo!()
-}
-
-/// Fetch and ensure that all dependencies are stored locally and return the
-/// paths to their contents. The returned map is guaranteed to have the same
-/// keys as [deps].
-/// contents. The returned map is guaranteed to have the same keys as [deps].
-pub async fn fetch<F: MoveFlavor>(
-    flavor: &F,
-    deps: DependencySet<PinnedDependencyInfo>,
-) -> PackageResult<DependencySet<FetchedDependency>> {
-    use DependencySet as DS;
-    use Pinned as P;
-    let mut result = DependencySet::new();
-    for (env, pkg, dep) in deps.into_iter() {
-        let fetched = PackagePath::new(dep.fetch().await?)?;
-        result.insert(env, pkg, FetchedDependency(dep.0.map(|_| fetched)))
-    }
-    Ok(result)
-}
-
-// TODO: unit tests
-#[cfg(test)]
-mod tests {}
